Move deferred handler tests to pytest
