+++ salt/payload.py
@@ -4,16 +4,12 @@ encrypted keys to general payload dynamics and packaging, these happen
 in here
 """
 
-
 import collections.abc
 import datetime
 import gc
-
-# import sys  # Use if sys is commented out below
 import logging
 
 import salt.loader.context
-import salt.log
 import salt.transport.frame
 import salt.utils.immutabletypes as immutabletypes
 import salt.utils.msgpack
