+++ salt/pillar/pillar_ldap.py
@@ -120,6 +120,7 @@ Result
         ]
     }
 """
+
 import logging
 import os
 
@@ -326,15 +327,9 @@ def ext_pillar(
         opts = salt.utils.yaml.safe_load(config_template) or {}
         opts["conf_file"] = config_file
     except Exception as err:  # pylint: disable=broad-except
-        import salt.log
-
-        msg = "pillar_ldap: error parsing configuration file: {} - {}".format(
-            config_file, err
+        log.warning(
+            "pillar_ldap: error parsing configuration file: %s - %s", config_file, err
         )
-        if salt.log.is_console_configured():
-            log.warning(msg)
-        else:
-            print(msg)
         return {}
     else:
         if not isinstance(opts, dict):
