+++ salt/utils/decorators/__init__.py
@@ -20,7 +20,6 @@ from salt.exceptions import (
     SaltConfigurationError,
     SaltInvocationError,
 )
-from salt.log import LOG_LEVELS
 
 IS_WINDOWS = False
 if getattr(sys, "getwindowsversion", False):
