+++ tests/filename_map.yml
@@ -9,9 +9,10 @@ setup.py:
   - pytests.scenarios.setup.test_install
 
 salt/_logging/(impl|handlers).py:
-  - unit.logging.test_deferred_stream_handler
   - pytests.integration._logging.test_jid_logging
   - pytests.integration._logging.test_multiple_processes_logging
+  - pytests.unit._logging.handlers.test_deferred_stream_handler
+
 
 salt/modules/(apkpkg|aptpkg|ebuildpkg|dpkg_lowpkg|freebsdpkg|mac_brew_pkg|mac_ports_pkg|openbsdpkg|opkg|pacmanpkg|pkgin|pkgng|pkg_resource|rpm_lowpkg|solarisipspkg|solarispkg|win_pkg|xbpspkg|yumpkg|zypperpkg)\.py:
   - pytests.unit.states.test_pkg
