+++ tests/pytests/unit/_logging/handlers/test_sentry_mod.py
