+++ tests/pytests/unit/transport/test_zeromq.py
@@ -9,7 +9,6 @@ import salt.config
 import salt.exceptions
 import salt.ext.tornado.gen
 import salt.ext.tornado.ioloop
-import salt.log.setup
 import salt.transport.zeromq
 import salt.utils.platform
 import salt.utils.process
